Add of validator services. Add of methods for contact editing.

+++ src/com/sirgiyenko/programm/model/Contact.java
@@ -24,7 +24,7 @@ public class Contact {
         this.name = name;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
